Fix phpstorm autocompletion
