+++ src/services/Http.php
@@ -4,7 +4,7 @@ namespace markhuot\craftpest\services;
 
 use craft\console\Application;
 use craft\helpers\App;
-use craft\web\Application as WebApplication;
+use markhuot\craftpest\web\Application as TestableWebApplication;
 use craft\web\Response;
 use craft\web\TemplateResponseBehavior;
 use craft\web\TemplateResponseFormatter;
@@ -12,6 +12,7 @@ use craft\web\UrlManager;
 use craft\web\User;
 use GuzzleHttp\Psr7\Message;
 use markhuot\craftpest\behaviors\TestableResponseBehavior;
+use markhuot\craftpest\web\TestableResponse;
 use markhuot\craftpest\web\Request;
 use Symfony\Component\Process\Process;
 use yii\base\Event;
@@ -26,20 +27,20 @@ class Http
      *
      * @see TestableResponseBehavior
      */
-    public function get(string $uri=null, $cookies = []): \markhuot\craftpest\web\Response
+    public function get(string $uri=null, $cookies = []): TestableResponse
     {
         $path = parse_url($uri, PHP_URL_PATH);
         $isCpRequest = str_starts_with(ltrim($path,'/'), 'admin');
         $request = Request::createGetRequestFromUri($uri, $isCpRequest);
         $request->cookies->fromArray($cookies);
 
-        /** @var WebApplication $craft */
+        /** @var TestableWebApplication $craft */
         $craft = \Craft::$app;
         $craft->getConfig()->getGeneral()->runQueueAutomatically = false;
 
         $craft->getView()->setTemplateMode($isCpRequest ? 'cp' : 'site');
         $craft->set('request', $request);
-        $craft->set('response', \markhuot\craftpest\web\Response::class);
+        $craft->set('response', \markhuot\craftpest\web\TestableResponse::class);
 
 
 
@@ -78,7 +79,7 @@ class Http
             // Support for status code checks
             if (is_a($e, HttpException::class)) {
 
-                $response = \Craft::createObject(\markhuot\craftpest\web\Response::class);
+                $response = \Craft::createObject(\markhuot\craftpest\web\TestableResponse::class);
                 $response->setStatusCode($e->statusCode);
 
                 // Error response
