+++ src/web/Application.php
@@ -2,10 +2,16 @@
 
 namespace markhuot\craftpest\web;
 
-/**
- * @property Request $request
- * @property Response $response
- */
-class Application extends \craft\web\Application {
+use yii\web\Response;
 
+/**
+ * @property Request          $request
+ * @property TestableResponse $response
+ */
+class Application extends \craft\web\Application
+{
+    public function handleRequest($request, bool $skipSpecialHandling = false): \markhuot\craftpest\web\TestableResponse
+    {
+        return parent::handleRequest($request, $skipSpecialHandling);
+    }
 }
