+++ stubs/init/Pest.php
@@ -13,7 +13,7 @@
 
 uses(
     markhuot\craftpest\test\TestCase::class,
-)->in(__DIR__);
+)->in('./');
 
 /*
 |--------------------------------------------------------------------------
