+++ tests/Pest.php
@@ -14,7 +14,7 @@
 uses(
     markhuot\craftpest\test\TestCase::class,
     markhuot\craftpest\test\RefreshesDatabase::class,
-)->in(__DIR__);
+)->in('./');
 
 /*
 |--------------------------------------------------------------------------
